# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Monopoly Session Server API
  version: 1.0.0
  description: |
    HTTP API for the standalone Monopoly session server (`StartSessionServer`).

    ## Session lifecycle

    1. `POST /sessions` to create a session with player names and colours.
    2. Open `GET /sessions/{id}/events` for a Server-Sent Events stream — the server
       pushes a full `ClientSessionSnapshot` after every accepted command.
    3. Submit game actions via `POST /sessions/{id}/command`.
    4. Poll `GET /sessions/{id}/snapshot` at any time for the current state.

    ## Commands

    All commands share a `type` discriminator field and a `sessionId` field.
    Most also require `actorPlayerId`. The server returns HTTP 200 with
    `accepted: true` on success, or HTTP 422 with rejection details on failure.

    ## CORS

    All responses include `Access-Control-Allow-Origin: *`.
    `OPTIONS` preflight requests return HTTP 204.

    ## Backward-compatible single-session endpoints

    When the server is started with player names as CLI arguments it also exposes
    `/command`, `/snapshot`, and `/events` without a session-ID path segment.
    These are omitted here; they use identical schemas.
servers:
  - url: http://localhost:8080
    description: Local dev server (default port)
...
paths:
...
  /sessions:
    get:
      tags: [meta]
      summary: List all active sessions
      operationId: listSessions
      responses:
        '200':
          description: Array of session summaries (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionSummary'

    post:
      tags: [meta]
      summary: Create a new game session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              names: [Jukka, Botti]
              colors: ['#E63946', '#2A9D8F']
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
        '400':
          description: Missing or empty `names` field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/command:
    post:
      tags: [session]
      summary: Submit a game command
      operationId: submitCommand
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCommand'
      responses:
        '200':
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
              example:
                accepted: true
                rejections: []
        '400':
          description: Unknown command type or malformed JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Command rejected by game rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
              example:
                accepted: false
                rejections:
                  - code: NOT_YOUR_TURN
                    message: It is not player-2's turn
  /sessions/{sessionId}/snapshot:
    get:
      tags: [session]
      summary: Get the current session snapshot (poll)
      operationId: getSnapshot
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Current game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSessionSnapshot'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{sessionId}/events:
    get:
      tags: [session]
      summary: Server-Sent Events stream of snapshot updates
      operationId: streamEvents
      description: |
        Opens a persistent SSE connection. The server immediately pushes the
        current snapshot as the first event, then pushes a new snapshot after
        every accepted command. A `": heartbeat"` comment is sent every 30 s
        to keep proxies from closing idle connections.

        Each event is a single `data:` line containing a full
        `ClientSessionSnapshot` JSON object followed by two newlines.

        ```
        data: {"sessionId":"...","version":5,"status":"IN_PROGRESS",...}

        ```
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: SSE stream (never terminates until client disconnects)
          content:
            text/event-stream:
              schema:
                type: string
                description: Newline-delimited SSE events carrying ClientSessionSnapshot JSON
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ──────────────────────────────────────────────────────────────────────────────
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Session ID returned by `POST /sessions`
  schemas:
...
    CreateSessionRequest:
      type: object
      required: [names]
      properties:
        names:
          type: array
          minItems: 2
          maxItems: 4
          items:
            type: string
          description: Player display names in seat order
          example: [Jukka, Botti]
        colors:
          type: array
          items:
            type: string
          description: |
            CSS hex colours for each seat (e.g. `#E63946`).
            If fewer colours than names are provided the server falls back to
            a default grey for the remaining seats.
          example: ['#E63946', '#2A9D8F', '#E9C46A', '#264653']
        seatKinds:
          type: array
          items:
            type: string
            enum: [HUMAN, BOT]
          description: |
            Seat kinds per player. Missing entries default to `HUMAN`.
            A `BOT` seat gets a `PureDomainBotDriver` attached automatically.
          example: [HUMAN, BOT]
        difficulties:
          type: array
          items:
            type: string
            enum: [NORMAL, STRONG]
          description: |
            Bot difficulty per seat. Only meaningful for `BOT` seats.
            Missing entries default to `NORMAL`.
          example: [NORMAL, STRONG]
    CreateSessionResponse:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    SessionSummary:
      type: object
      required: [sessionId, playerNames, status]
      properties:
        sessionId:
          type: string
          format: uuid
        playerNames:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/SessionStatus'

    # ── Command result ───────────────────────────────────────────────────────
    CommandResult:
      type: object
      required: [accepted, rejections]
      properties:
        accepted:
          type: boolean
        rejections:
          type: array
          items:
            $ref: '#/components/schemas/CommandRejection'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    # ── Snapshot ─────────────────────────────────────────────────────────────
    ClientSessionSnapshot:
      type: object
      required: [sessionId, version, status, viewAvailable]
      properties:
        sessionId:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
          description: Monotonically increasing state version
        status:
          $ref: '#/components/schemas/SessionStatus'
        viewAvailable:
          type: boolean
          description: True when the session host has a live game view attached
        state:
          $ref: '#/components/schemas/SessionState'
...
    SessionStatus:
      type: string
      enum: [IN_PROGRESS, GAME_OVER]
...
    SessionCommand:
      discriminator:
        propertyName: type
        mapping:
          RollDice:                         '#/components/schemas/RollDiceCommand'
          EndTurn:                          '#/components/schemas/EndTurnCommand'
          BuyProperty:                      '#/components/schemas/BuyPropertyCommand'
          DeclineProperty:                  '#/components/schemas/DeclinePropertyCommand'
          PayDebt:                          '#/components/schemas/PayDebtCommand'
          MortgagePropertyForDebt:          '#/components/schemas/MortgagePropertyForDebtCommand'
          SellBuildingForDebt:              '#/components/schemas/SellBuildingForDebtCommand'
          SellBuildingRoundsAcrossSetForDebt: '#/components/schemas/SellBuildingRoundsAcrossSetForDebtCommand'
          DeclareBankruptcy:                '#/components/schemas/DeclareBankruptcyCommand'
          PassAuction:                      '#/components/schemas/PassAuctionCommand'
          PlaceAuctionBid:                  '#/components/schemas/PlaceAuctionBidCommand'
          FinishAuctionResolution:          '#/components/schemas/FinishAuctionResolutionCommand'
          OpenTrade:                        '#/components/schemas/OpenTradeCommand'
          SubmitTradeOffer:                 '#/components/schemas/SubmitTradeOfferCommand'
          AcceptTrade:                      '#/components/schemas/AcceptTradeCommand'
          CounterTrade:                     '#/components/schemas/CounterTradeCommand'
          DeclineTrade:                     '#/components/schemas/DeclineTradeCommand'
          CancelTrade:                      '#/components/schemas/CancelTradeCommand'
          EditTradeOffer:                   '#/components/schemas/EditTradeOfferCommand'
          BuyBuildingRound:                 '#/components/schemas/BuyBuildingRoundCommand'
          ToggleMortgage:                   '#/components/schemas/ToggleMortgageCommand'
          RefreshSessionView:               '#/components/schemas/RefreshSessionViewCommand'
      oneOf:
        - $ref: '#/components/schemas/RollDiceCommand'
        - $ref: '#/components/schemas/EndTurnCommand'
        - $ref: '#/components/schemas/BuyPropertyCommand'
        - $ref: '#/components/schemas/DeclinePropertyCommand'
        - $ref: '#/components/schemas/PayDebtCommand'
        - $ref: '#/components/schemas/MortgagePropertyForDebtCommand'
        - $ref: '#/components/schemas/SellBuildingForDebtCommand'
        - $ref: '#/components/schemas/SellBuildingRoundsAcrossSetForDebtCommand'
        - $ref: '#/components/schemas/DeclareBankruptcyCommand'
        - $ref: '#/components/schemas/PassAuctionCommand'
        - $ref: '#/components/schemas/PlaceAuctionBidCommand'
        - $ref: '#/components/schemas/FinishAuctionResolutionCommand'
        - $ref: '#/components/schemas/OpenTradeCommand'
        - $ref: '#/components/schemas/SubmitTradeOfferCommand'
        - $ref: '#/components/schemas/AcceptTradeCommand'
        - $ref: '#/components/schemas/CounterTradeCommand'
        - $ref: '#/components/schemas/DeclineTradeCommand'
        - $ref: '#/components/schemas/CancelTradeCommand'
        - $ref: '#/components/schemas/EditTradeOfferCommand'
        - $ref: '#/components/schemas/BuyBuildingRoundCommand'
        - $ref: '#/components/schemas/ToggleMortgageCommand'
        - $ref: '#/components/schemas/RefreshSessionViewCommand'

    # Base shared by most commands
    BaseCommand:
      type: object
      required: [type, sessionId]
      properties:
        type:
          type: string
          description: Command type discriminator
        sessionId:
          type: string
          format: uuid
...
    RollDiceCommand:
      title: RollDice
      description: Roll the dice to start the active player's turn. Only valid during `WAITING_FOR_ROLL`.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          properties:
            type:
              type: string
              enum: [RollDice]
      example:
        type: RollDice
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1
    EndTurnCommand:
      title: EndTurn
      description: End the active player's turn. Only valid during `WAITING_FOR_END_TURN`.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          properties:
            type:
              type: string
              enum: [EndTurn]
      example:
        type: EndTurn
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1

    # ── Property purchase ──────────────────────────────────────────────────
    BuyPropertyCommand:
      title: BuyProperty
      description: |
        Purchase the property the active player has landed on.
        Requires `decisionId` from `state.pendingDecision.decisionId` and
        `propertyId` from the decision payload.
        Only valid during `WAITING_FOR_DECISION`.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [decisionId, propertyId]
          properties:
            type:
              type: string
              enum: [BuyProperty]
            decisionId:
              type: string
              format: uuid
              description: From `state.pendingDecision.decisionId`
            propertyId:
              type: string
              example: B1
      example:
        type: BuyProperty
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1
        decisionId: a1b2c3d4-0000-0000-0000-000000000000
        propertyId: B1
    DeclinePropertyCommand:
      title: DeclineProperty
      description: |
        Decline to purchase the property and send it to auction.
        Requires the same `decisionId` and `propertyId` as `BuyProperty`.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [decisionId, propertyId]
          properties:
            type:
              type: string
              enum: [DeclineProperty]
            decisionId:
              type: string
              format: uuid
            propertyId:
              type: string
      example:
        type: DeclineProperty
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1
        decisionId: a1b2c3d4-0000-0000-0000-000000000000
        propertyId: B1

    # ── Debt resolution ────────────────────────────────────────────────────
    PayDebtCommand:
      title: PayDebt
      description: |
        Pay the outstanding debt in full from the player's cash.
        Only valid during `RESOLVING_DEBT` when the player has sufficient cash.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [debtId]
          properties:
            type:
              type: string
              enum: [PayDebt]
            debtId:
              type: string
              format: uuid
              description: From `state.activeDebt.debtId`
      example:
        type: PayDebt
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-2
        debtId: d0000000-0000-0000-0000-000000000001
    MortgagePropertyForDebtCommand:
      title: MortgagePropertyForDebt
      description: Mortgage one of the debtor's properties to raise cash during debt resolution.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [debtId, propertyId]
          properties:
            type:
              type: string
              enum: [MortgagePropertyForDebt]
            debtId:
              type: string
              format: uuid
            propertyId:
              type: string
              example: B1
      example:
        type: MortgagePropertyForDebt
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-2
        debtId: d0000000-0000-0000-0000-000000000001
        propertyId: B1
    SellBuildingForDebtCommand:
      title: SellBuildingForDebt
      description: Sell one or more buildings on a property to raise cash during debt resolution.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [debtId, propertyId, count]
          properties:
            type:
              type: string
              enum: [SellBuildingForDebt]
            debtId:
              type: string
              format: uuid
            propertyId:
              type: string
            count:
              type: integer
              minimum: 1
              description: Number of buildings to sell (1 sells one house or hotel)
      example:
        type: SellBuildingForDebt
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-2
        debtId: d0000000-0000-0000-0000-000000000001
        propertyId: O1
        count: 1
    SellBuildingRoundsAcrossSetForDebtCommand:
      title: SellBuildingRoundsAcrossSetForDebt
      description: |
        Sell buildings evenly across an entire colour group (one round = one building
        from every property in the group). Useful for rapidly liquidating a developed set.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [debtId, propertyId, rounds]
          properties:
            type:
              type: string
              enum: [SellBuildingRoundsAcrossSetForDebt]
            debtId:
              type: string
              format: uuid
            propertyId:
              type: string
              description: Any property in the colour group
            rounds:
              type: integer
              minimum: 1
    DeclareBankruptcyCommand:
      title: DeclareBankruptcy
      description: Declare bankruptcy, eliminating the player from the game.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [debtId]
          properties:
            type:
              type: string
              enum: [DeclareBankruptcy]
            debtId:
              type: string
              format: uuid
      example:
        type: DeclareBankruptcy
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-2
        debtId: d0000000-0000-0000-0000-000000000001

    # ── Auction ────────────────────────────────────────────────────────────
    PassAuctionCommand:
      title: PassAuction
      description: Pass (skip bidding) for the current auction round.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [auctionId]
          properties:
            type:
              type: string
              enum: [PassAuction]
            auctionId:
              type: string
              format: uuid
              description: From `state.auctionState.auctionId`
      example:
        type: PassAuction
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1
        auctionId: e0000000-0000-0000-0000-000000000002
    PlaceAuctionBidCommand:
      title: PlaceAuctionBid
      description: |
        Place a bid in the current auction.
        `amount` must be ≥ `state.auctionState.minimumNextBid`.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [auctionId, amount]
          properties:
            type:
              type: string
              enum: [PlaceAuctionBid]
            auctionId:
              type: string
              format: uuid
            amount:
              type: integer
              minimum: 1
      example:
        type: PlaceAuctionBid
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1
        auctionId: e0000000-0000-0000-0000-000000000002
        amount: 120
    FinishAuctionResolutionCommand:
      title: FinishAuctionResolution
      description: Finalise a resolved auction and transfer the property to the winner.
      allOf:
        - $ref: '#/components/schemas/BaseCommand'
        - type: object
          required: [auctionId]
          properties:
            type:
              type: string
              enum: [FinishAuctionResolution]
            auctionId:
              type: string
              format: uuid

    # ── Trade ──────────────────────────────────────────────────────────────
    OpenTradeCommand:
      title: OpenTrade
      description: Initiate a trade negotiation with another player.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [recipientPlayerId]
          properties:
            type:
              type: string
              enum: [OpenTrade]
            recipientPlayerId:
              type: string
    SubmitTradeOfferCommand:
      title: SubmitTradeOffer
      description: Submit the current draft offer to the other player for review.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [tradeId]
          properties:
            type:
              type: string
              enum: [SubmitTradeOffer]
            tradeId:
              type: string
              format: uuid
    AcceptTradeCommand:
      title: AcceptTrade
      description: Accept the submitted trade offer and execute the exchange.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [tradeId]
          properties:
            type:
              type: string
              enum: [AcceptTrade]
            tradeId:
              type: string
              format: uuid
    CounterTradeCommand:
      title: CounterTrade
      description: Reject the offer and send it back for editing by the other side.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [tradeId]
          properties:
            type:
              type: string
              enum: [CounterTrade]
            tradeId:
              type: string
              format: uuid
    DeclineTradeCommand:
      title: DeclineTrade
      description: Decline the submitted offer and end the trade.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [tradeId]
          properties:
            type:
              type: string
              enum: [DeclineTrade]
            tradeId:
              type: string
              format: uuid
    CancelTradeCommand:
      title: CancelTrade
      description: Cancel an open trade before an offer has been submitted.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [tradeId]
          properties:
            type:
              type: string
              enum: [CancelTrade]
            tradeId:
              type: string
              format: uuid
    EditTradeOfferCommand:
      title: EditTradeOffer
      description: |
        Apply a partial update to the active trade offer.
        Only the fields present in `patch` are changed; omitted fields are left as-is.
        Common patch fields: `offeredCash` (int), `addPropertyId` (string),
        `removePropertyId` (string).
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [tradeId, patch]
          properties:
            type:
              type: string
              enum: [EditTradeOffer]
            tradeId:
              type: string
              format: uuid
            patch:
              type: object
              example:
                offeredCash: 200
                addPropertyId: B1

    # ── Building / mortgage ────────────────────────────────────────────────
    BuyBuildingRoundCommand:
      title: BuyBuildingRound
      description: |
        Buy one building on every property in the colour group (one round).
        `propertyId` can be any property in the target group.
      allOf:
        - $ref: '#/components/schemas/BaseActorCommand'
        - type: object
          required: [propertyId]
          properties:
            type:
              type: string
              enum: [BuyBuildingRound]
            propertyId:
              type: string
      example:
        type: BuyBuildingRound
        sessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        actorPlayerId: player-1
        propertyId: O1
...